refactor: move catalog validation into CLI
